Add Visual Lint to Compose preview UI check mode
This means that Visual Lint checks will be run in addition to ATF checks
in UI check mode for the Compose preview. Just as in the case of Layout
Validation, the errors coming from Visual Lint will be grouped if the
same error on the same component appears in several previews.

Bug: 274150043
Test: test added
Change-Id: I2252fa83bdb36112771fbb4c6ed5d8426dca2ef1

GitOrigin-RevId: ea9cca19c7b4a38455e18679726d912103b15da8