Add a 'topic' attribute to the rendering process
This new attribute is intended to be used for specifying the context
or tool in which a given render is happening. By default, if not
specified, a 'Not specified' topic is used.

Also, two main changes are introduced in the cancellation mechanism
provided by RenderExecutor:
1- A new cancellation mechanism by topic is being added
2- The render action that is running could be interrupted if wanted
by using the parameter 'mayInterruptIfRunning'.

In this CL Visual Lint is the only tool changed to use its own
topic because it is the only one using the existing cancellation
mechanism. However, no behavioral changes should happen.

Follow up CLs will make other tools use the new RenderingTopic

Bug: 268590982
Bug: 279604596
Test: RenderExecutorTest
Change-Id: Ib77e19ff2863365efd7599ac349b31d4d9e0120f

GitOrigin-RevId: ce23d9bceb255f8dc34ed44bdd6659563c905f32