Take the minimum 1st column width into account when resizing
The column resize in the properties panel did not take the minimum
width of the first column into account when determining the
nameColumnFraction.

As a result the user would set the hotspot for resizing at a
position that did not match the visual column divider in the table
thus making it hard to resize the column back.

Fixes: 279926340
Test: ColumnFractionChangeHandlerTest
Change-Id: I80ae766384669098d716ae83cf6dc47a4d36d034

GitOrigin-RevId: b846c70ce7fcee4dfd7329e399b7373b8f6829cf