Delay IssueModel initialization in VisualLintService
We're passing the VisualLintService as the IssueModel parent disposable
when setting it as a property of the service. The problem is: if there
is an error in the VisualLintService initialization, an object will not
be properly passed to IssueModel, and it might never be disposed, which
results in a Project leak. This change moves the initialization of the
IssueModel property to be the last thing that's done in the constructor.

Bug: 279929471
Test: designer and compose-designer tests
Change-Id: I9f69647951f35da03c1fc94ccd04ceb0b0b40ef3

GitOrigin-RevId: 3bfc2f215b79f5ffe9f062d6bb3510385737cc2a