Reenable testUseScope
I suspect that this flakiness may have been addressed with ag/22815207.
Prior to that change, the androidTest directories added in this test
were technically not part of the project, due to not being under any
source root. Because of that, I think some of the guarantees about class
indexes being up to date when they're queried may not have been valid,
and so on some occasions the class was not available yet.

I'm re-enabling this test following the above fix to try out this
theory. (I've tried and have not been able to reproduce the failure
either with or without that fix locally.) I'll monitor to see if this
recurs. Also adding some explicit assertions to help confirm exactly
what's failing.

Bug: 279233700
Test: Test fix
Change-Id: I317f8e4173b392b242d4a2943b9f5325b2a6b4ef

GitOrigin-RevId: 9859f3aada36845bd2260fe8295965ba4bdea901