Adapt IssueChecker to Gradle TAPI 8.0
The Tooling API used by the platform has increased in version to 8.0,
which handles Gradle distributions managed by GradleWrapper
differently: the installation process now deletes the .zip file if
unzip is successful, even if the overall installation is not -- which
it tracks using a state file with name ...zip.ok.  Adapt the Gradle
Issue Checker to check for the existence of the .ok file if the .zip
file is not present, reasoning that if we have an .ok file but a
failure message, probably that .ok file is wrong.

Fixes: 280506909
Test: re-enabled, added
Change-Id: I540bdb3ae4a7c71e2ccdf3e513041d902252bcfd

GitOrigin-RevId: 7769db25322f11620203fa58cc8f2d67dbf05b7f