Don't hold a strong Project reference to prevent leaks
We were directly storing a Project value in ComposePreviewRepresentation
and this was causing a leak of Project reference in tests when there is
a failure that doesn't cleanup the ComposePreviewRepresentation object.

For example, RenderErrorTest and AccessibilityModelUpdaterTest call the
test utility method activateAndWaitForRender, which has a timeout of 30s
and will throw a TimeoutCancellationException if this time is exceeded.

In order to prevent this leak from happening, we can access the project
property through the psiFilePointer value we're already setting.

Bug: 272048512
Bug: 279558854
Test: compose-designer tests
Change-Id: I76264cf6c6f8247cabd6ba3f74020268ee65db7b

GitOrigin-RevId: 1d1a589700ddd77bc3266d632881c5db4e557bb7