Ensure file updates are processed before leaks check
PerFileElementTypeStubModificationTracker added recently keeps a queue
of file operations, items of this queue contain references to the
Project. Normally this queue is processed reasonably fast and items for
disposed projects are just skipped.
However, in test, if there are heavy file operations and not enough time
to process them in the end of the test it can happen that leak  checker
runs before all the file updates are processed resulting in a disposed
project leak detection.
Fix this by force-running this queue processing before running the leak
checker.
This fixes flakiness of ProjectUpgradeTests_* and possibly
some other tests.

Bug: 279393325
Test: Fixing flaky tests
Change-Id: Ib007c3d8d4d3ba80699bc7cc70f666af9bf3f1f6

GitOrigin-RevId: b8a9d3c60622c6745a5595f8c98cb4211a7a8587