[AUA] Fix crash in upgrading plugins
It's possible to have dependency definitions that, for whatever
reason, not only have unresolvable versions but have those versions be
entirely missing from the dependency's model.  That state is probably
not entirely valid (though there are indications in the referenced bug
that it is somehow working for some people) but we should not allow
that state to cause the upgrader for GradlePlugins to crash.

How does it crash?  An innocuous-looking toString() on a
GradlePropertyModel can return `null` if the model is missing, which
we can cause to happen if we have a map-style dependency (as naturally
comes from using Version Catalogs with missing fields).

Defend against this by checking that the model's ValueType is STRING,
which will be the case if we resolve it to a String value (whereupon
toString() will return that String).  If we don't resolve it to a
String value, it is correct for us to ignore that dependency as we
cannot know what to do with it.

Bug: 278102583
Test: added
Change-Id: I0ef24565ec738fe0f077cce7709ab35dcd1185a5

GitOrigin-RevId: 3f51ffd8d54c858d9b142cb9249e98defd62b664