Handle TimeoutCancellationException in Compose Preview tests
When activateAndWaitForRender exceeds the timeout set in Compose Preview
tests, the method throws a TimeoutCancellationException that we were not
handling. When that happens, the test doesn´t finish gracefully and the
the leak check rule will detect Project leaks because the objects that
hold a reference to a Project were not disposed properly.

We need to handle the TimeoutCancellationException and properly fail the
test, so disposal happens correctly and the entire target is not
compromised by a single test failure.

Bug: 279899581
Bug: 272048512
Test: compose-designer tests
Change-Id: I2bce71e68dc67db3ca4540137204a66190e8c791

GitOrigin-RevId: 5e1af34e8e455121f3f55df94993c589e877ee47