Add metric tracking for track group mouse over
The MOUSE_OVER type has been added as a new
TrackGroupActionType under the AdtUiTrackGroupMetadata
metric event.

This new metric type will allow us to track the usage of
the power profiler, but was implemented in a generalized
way to track usage of any track group. The metric is tracked
everytime the track group content is moused over.

Bug: 254116554
Test: manual
Change-Id: I37a426e712dfd706dd693720252a32261a19c2d0

GitOrigin-RevId: b78c2a8ab6bac49438a5130fe3f56cd835507d67