Disable flaky test in editors target
//tools/adt/idea/android:intellij.android.core.tests_tests__editors
has been flaking on Windows recently. The entire target times out even
though when it passes it does so in 2 minutes or less, pointing to a
probable deadlock.

Locally I've reproed this and identified a single test that occasionally
locks. This change is @Ignore'ing that test for now to reduce flakiness,
but I'll keep the bug open to track re-enabling it.

(Note my local repro was on Linux, not Windows - but since I've
identified this as a deadlock regardless, it makes sense to disable
until it can be fixed. I'll monitor the Windows post-submit to see if
there are further failures.)

Bug: 281109986
Test: test change
Change-Id: I83eb291adc822a35dc7480fe24bdff318724e456

GitOrigin-RevId: f61b0392e073df8914f37c6f85af6de5a1f6f2fa