Address flakiness in AndroidResourceRenameTest
Similar speculative fix as in ag/23149979.

I went through recent test target failures and tried to collect the
various methods that were failing (there were a lot). They had similar
characteristics to what I tried to address in ag/23149979, so I'm trying
a similar fix here.

Worst case, these changes shouldn't affect anything. So I'm wanting to
get this change in so that I can see if there's any effect on the
flakiness in automated builds.

(Note that one test method was disabled for b/281863312, and I haven't
reenabled that here. But I've seen many other failing methods, so this
change is trying to address a number of those before I might reenable
that one.)

Bug: 281863312
Test: n/a (test change)
Change-Id: Ic65a3ddbbc4b38ad6bd6954ad6025420704479d5

GitOrigin-RevId: 756d6f555793a14805ea941a2c098ba880f01dfc