Add diagnostics flag for the compose inspector
Recent external bug reports have been hard to diagnose because the
developer is unable to share the project, and are also filtering
the log entries for names that may reveal stuff about their
application.

As a start I added some diagnostics printouts for the compose
inspector. We can ask a developer to turn this on in their custom
vm options.

Bug: N/A
Test: Existing tests should not fail.
Change-Id: I95cc2667f03a1138344758c6d09ac7825399e283

GitOrigin-RevId: b32e4c8000c4214f9397b202ce099ca0d3c5ce6f