Remove adding screenshot padding to fit 384x384 dimensions
...as the dimensions are not checked during playstore submission and are
only just indicative on DAC.

Furthermore, we received some further clarification stating that the
edge of the UI should extend to the edge of the screenshot. This means
that adding padding to the screenshot rendered the screenshot
incompatible with the playstore requirements.

This CL also changes the behaviour to include the "play store" option
only if the screenshot image has a 1:1 ratio as we cannot make it
play store compliant by adding padding.

Bug: 235061954
Test: update ScreenshotViewerTest and DeviceScreenshotPostprocessorTest
Change-Id: Id515e1cc2046de3edd84e38fb77e37b6fd5b9255

GitOrigin-RevId: 8d3297cf9940bce428f4bb3e933ab537b914de32