Android View: rearrange source type nodes setup
Add a new source type that explicitly states in the name it is for
mixed kotlin and java sources. An intersection of kotlin and
java source sets will be listed under it.
Java and Kotlin types will be left with the remaining non-intersecting
sub-sets if they are configured.

This way it will be clearly visible where go the java files, the kotlin
files or both together.

One caveat is that kotlin sourcesets are reported even in the case when
no kotlin is configure for the project. Because of this we need to
check in the facets if kotlin is configured and just add  all roots from
the intersection to the Java part if Kotlin is missing.

Bug: 273905575
Test: updated snapshots
Change-Id: I18b84df4a78ba57495c6fcebfa60d39b952d65ff

GitOrigin-RevId: 944ba856af4a85f0008f3755546360d33eaf8991