Convert AndroidFindStyleApplicationsTest
Converting to use AndroidProjectRule instead of AndroidTestCase, and
changing to Kotlin at the same time.

b/281709998 is technically for a different test, but this change is
related since this test starts failing as soon as the other test is
moved out of the suite. Too bad the tests aren't already hermetic...

Bug: 281709998
Test: AndroidFindStyleApplicationsTest
Change-Id: I19a017f2de95c5e95ee8a530911776c4c8365bc3

GitOrigin-RevId: 76849191c5f381c734afce2f69e8d3b502856511