Given an android compilation support having multiple valid source set
The KMP populates the KotlinMPPGradleModel with some wrong Android
compilations and source sets. This happens because the KMP does
neither understand Android variants nor the single-variant sync.
Because of that, we end up removing the wrong compilation and
merging the non-needed source sets. Until now our logic relayed
on having just one possible valid android source set
[androidMain, androidAndroidTest, androidTest] per android
compilation. This isn't exactly true when users have defined for
their androidAndroidTest source set a dependency with androidTest.

val androidAndroidTest by getting {
   dependsOn(androidTest)
}

For this reason, we added support to be able to cover this scenario.

Bug: 282068272
Test: testKotlinMultiplatform_multiple_source_set_per_android_compilation
Change-Id: I8656d559de8cfeb6df221d9f7fc45f6052e010d0

GitOrigin-RevId: 7c248b2f7de5f29698c90225fd319ce98b21abd9