Remove custom Android component safe deleter
This deleter was added long ago to solve
https://youtrack.jetbrains.com/issue/IDEA-25589. From reading that bug,
the desired behavior was different than what seems to be the norm for
safe delete, at least as it now exists 12 years later. That bug wanted
to automatically delete the manifest entry for an activity (or other
component) when the user tries to safe delete an Activity class, as long
as there are no other code usages.

Current behavior of other components implies a different behavior is
desired. To be more in line with other parts of the editor, the manifest
usage should indeed be called out as part of Safe Delete, and the user
should be given a choice whether they want to continue or not.

The behavior I described as what I think we should have now is actually
what's happening in the product; and
`AndroidComponentSafeDeleteProcessor` is basically dead code. This can
be seen with b/110489091 and the corresponding ag/14705889. The addition
of the Kotlin plugin made the original behavior stop working. This was
because the Kotlin safe delete processor was registered before this
custom Android component processor, and was "stealing" processing of the
deletion. In doing so, it inadvertently gave us the newer behavior we
want.

The fix to that breakage actually changed the test code to reflect the
newer behavior, rather than restoring what the processor used to do. I
think it was probably just missed that the processor was no longer
running.

Because of all of the above, we want to keep the safe delete test to
keep validating behavior, but we no longer need a custom safe delete
processor to make this happen.

Fixes: 282812469
Test: AndroidSafeDeleteTest
Change-Id: Ib8bd1fc33e4c84932a23cab517b7f2e8d8108b3a

GitOrigin-RevId: 67e913749a6af605efeb7faee92dc01c89207ea7