Rewrite tests that exercise only platform code to be more flexible
The test methods string3{Java,Kotlin}() are actually testing that
we properly fall back to platform docs for cases we do not handle.
Instead of checking the HTML exactly (which we were doing before,
but broke when IJ changed the order in which things were displayed),
roughly check for the presence of what we think should be included.

Fixes: 243077207
Test: AndroidJavaDocRendererTest.java
Change-Id: Id6a4922a10e444302b96eaed52025b0be8c5d837

GitOrigin-RevId: bc6ea45bbdcd308a594cfd0caad1625fce77eb44