Migrate JDK integration tests from JDK_17 to JDK_EMBEDDED
The JDK_EMBEDDED represents the latest studio embedded JDK version,
that right now points to JDK_17. This will change in the future and
most of our tests are using explicitly JDK_17 as the latest one,
something that will make it difficult for the next JDK migration.

For this reason, all the tests that don't strictly require the JDK 17
have been migrated to use the latest Embedded JDK.

Bug: 273937598
Test: Existing
Change-Id: Ia0db5701db9ebae5954ffda9dd95387e61854192

GitOrigin-RevId: 65f160f25863f8746e78b5e12d5582ce778cb917