Move MoveClass* methods
Interestingly, the test data for these classes wasn't used, and the old
version would do the rename but didn't actually verify the results. It
looks like it's been that way since the test was added. At any rate,
I've added verification that the rename actually did what it was
supposed to, not just that it completed without throwing.

Bug: 281863312
Test: AndroidMoveActivityTest
Change-Id: I73123d8c904c353ec852684781822d35a9e6007a

GitOrigin-RevId: fc3640cd0e1d0951e0ccdf50ca29abe2dfbbbf4b