Replace Deprecated Library with New DSL
`com.intellij.ui.layout` is deprecated and scheduled for removal.

The new code was written from scratch but based on the old code.

Note that the old code was split into 2 versions of the dialog:
1. LogcatFormatDialog - For custom formats
2. LogcatFormatPresetsDialog - For presets

With a base class LogcatFormatDialogBase.

The #1 option was never enabled and never actually a requirement, so I
removed it and consolidated the code into a single class.

Bug: NA
Test: Updated
Change-Id: I01769cacfcb51500d4e2540119884803fe7bd667

GitOrigin-RevId: afcb318be37993586110a5da50c458916bb5eaca