Add sync back to generating template golden files
TemplateDiffTest now has 3 modes: DIFFING, VALIDATING, and GENERATING.
When generating golden files, VALIDATING and GENERATING are run
sequentially, with VALIDATING using AndroidGradleProjectRule for Gradle
sync and later would be used for building and Linting. GENERATING is run
without sync so the final output files are the same as DIFFING, which
also does not sync.

Bug: 272145738
Test: this
Change-Id: I44e4b81f55427efcc675639bf6c1cc3e574517de

GitOrigin-RevId: 9425070760643d304cbc157ba4f8ac1505a233b5