Reconcile screen rotation from Layout Inspector and Running Devices
Both the rendering from Layout Inspector and Running Devices can be
rotated, independently of each other.

This change changes the way the rendering from Layout Inspector
is overlaid to the display from Running Devices so that they
always match. To prevent situations where, for example, one is
horizontal and the other is vertical.

Test: LayoutInspectorRendererTest
Bug: 265150325
Change-Id: I0bc98365331f3c9d70a681bcb8ca08eb94383cf3

GitOrigin-RevId: 9a291f7a37049de9c7fdbe794e85f0ad924f8719