Update ComposableIconProvider to not duplicate super class
Now that https://github.com/JetBrains/intellij-community/pull/2204 has
come into the code base, we no longer need this provider to return the
super class's value. When it returns null, the calling code will
continue on to the next provider and try it instead, which will have the
same effect. This new structure is desirable, because it allows another
extension to insert itself as well, which wouldn't have otherwise been
possible.

Fixes: 255608484
Test: ComposableIconProviderTest (covered by existing tests)
Change-Id: I5f49ce99f95631a5382e4266342a9daf2bb2732f

GitOrigin-RevId: b5d4964b80cab3b1624badd56112da47bc9a4517