Avoid ModuleClassLoader references from the disposer
Since the ModuleClassLoader has its own separate dispose mechanism using
release, Disposer references can cause unintended leaks when
Disposer.dispose is not correctly called in addition to release.
Disposer.dispose was being called from a number of places making the
ownership of the ModuleClassLoader not clear and leaving the
responsability to call .dispose in a few places.

This removes the top level Disposable to avoid the temptation of
registering and leaving the only release responsibility to the
ModuleClassLoaderManager.

Test: StudioModuleClassLoaderManagerTest
Fixes: 264921829
Change-Id: I2a389c25950bd7e6de4e3398259a35e8fa1bfe97

GitOrigin-RevId: f2f8bdc183d7b7ebd2494f902a4beebefe96c588