Fix UpdateTest
After restarting Android Studio, the test tries re-attaching to the
Android Studio process. This was taking longer than 30 seconds. As a
result, I increased the timeout to 120 seconds. This brought the failure
rate from ~95% to 0%.

Strangely, the test always succeeded for me locally even with the
original timeout, although I couldn't run anywhere near the same number
of times locally.

Fixes: 284193142
Test: UpdateTest
Change-Id: I6351d2db0b21d53b3b255db2b0ca806d936c7b93

GitOrigin-RevId: 0d881cb2e4b4b8da5cb05d597fc2e0cfced7a373