Update SymbolSourceTest to use Truth across all its tests
We are seeing timeout issues from the tests in SymbolSourceTest, it is
pretty consistent but the logic being tested is pretty simple and not
time related, so it is unclear what's causing the timeouts.

However, looking at the tests, there is a mixture of Assert and Truth.
This change goes through the tests and converts them to all use Truth.

Bug: 283765759
Test: Ran the tests that were changed.
Change-Id: I87aabc0276d1a08c3b4fffe0bc975d014823f231

GitOrigin-RevId: db31435c8eb021d9e0616ad43be9d2402abe4b3f