Don't translate rendering in the model
Before this change, translation happened in the model, to always center
the rendering. As a consequence, the x, y coordinates of the root view
were always (0,0), even if the real x, y on the device were not (0,0).

With this change, view coordinates in the rendering always match the
coordinate of the views on the device.
For example if on the device the view x, y are (0, 50), the view will
also appear 50 pixels from the top of the screen in the LI rendering.

This makes it easier to work with the rendering and to overlay the
device screen to another image, like the screen image from Running
Devices.

Test: existing
Bug: 265150325
Change-Id: I075b58bc1ca819d12e3e137876c7f83ba2703358

GitOrigin-RevId: e009f4b16783b76b26a37c807116b02517ee83f7