Make PreviewRefreshManager responsible of cancellations
Before this change, the PreviewRefreshManager was responsible for
detecting the need of cancelling a refresh, but the cancellation
was being delegated to the requests. This was causing some race
conditions in PreviewRefreshManagerTest, and probably also in
ComposePreviewRefreshRequest.

This change removes the 'cancel' method from PreviewRefreshRequest
and makes the generic manager responsible for cancelling requests
in order to keep the code simple and avoid similar problems in
any other tool that may use a PreviewRefreshManager in the future.

About PreviewRefreshManagerTest, before this change, 16 executions
failed out of 500, and after this change, no failure happened when
executing with --runs_per_test=1000.

Fixes: 284479175
Test: PreviewRefreshManagerTest and ComposePreviewRefreshManagerTest
Change-Id: I2ba4edd1e263674214ebc784f06a9027a8f6f4ed

GitOrigin-RevId: 5fdbfcd8a457035076f48a7c867851c41f78a9b4