Do not run image based ATF checks in background visual lint
The renderings done in background visual lint are of low quality for
performance reason, but that low quality is causing mistakes in the
image-based ATF checks. To avoid that, let's only run ATF checks when
the rendering is of high enough quality, that is in cases where we
actually use the rendering for display (e.g. in Layout Validation).

Bug: N/A
Test: tests updated
Change-Id: I3977bf2089838ce5c161bb169f3e2fa475bb1fb7

GitOrigin-RevId: 603940ce8c2049440af13dc9db61cb02df914d0d