Add Kotlin Script to LintIdeGradleDetectorTest
Port almost all the existing tests to Kotlin Script Dsl;
some (assignment to minSdkVersion and handling of `apply plugin:` have
no equivalent in Kotlin Script, so exit early from those tests for the
Kotlin Script instantiation of the parameterized test.

We need to mask the ProblemHighlightingFilter for this test because
the Kotlin IDE plugin registers a filter turning off highlighting for
Kotlin Script files with no configuration, but such a configuration
can only be obtained by doing a much fuller project import than is
provided by the AndroidTestBase fixture.

Bug: 284086015
Test: this is more tests
Change-Id: I9ae130bbfe1b1edd5037920d3ae2927d902cc376

GitOrigin-RevId: e984153e3a092a526b96e2ad1646280d3c0a01a5