Break down the time measurements further
This adds measurements for:
- Configuration time
- Time spent before android models start building
- Time spent building android models
- Time spent after android models finish building

Studio flag that outputs sync stats now accepts a directory
to store the output, so it can be processed by tests. This is
used to capture configuration finished timestamp now, it will
also be used to capture individual phases of Android model building later.

The former "before_sync" measurement point roughly corresponded to when the
project import action was instantiated (which was a rough approximation of
configuration finished, but not precise). Now that we measure that more
precisely, it's replaced with the android model building start point.

Bug: 241546559
Test: Test change
Change-Id: I96a98deb276af5b63923c359135dae0794bc94d9

GitOrigin-RevId: 6b66eb7d2aaa5b07d4ed231e8d94e5c5926cca0f