Remove checkConfiguration in BaseAction
checkConfiguration can be potentially long operation and because update
of ApplyChangesAction was moved to BGT in
I15ff410f245b1ad294786b7697142840eae2b8d1 it has conflict with UI thread
 over read lock.

We didn't have checkConfiguration in BaseAction before Flamingo, it was
introduced to make AC and ACC more accurate than they used to be. But
benefits it provides are outweighed by the UI freeze.

This commit reverts back to the old system where AC and ACC are not
blocked if the configuration is invalid. Instead, when the user tries to
 execute them, a popup with an error message is displayed.

Fixes: 284169566
Test: make configuration invalid and try to AC, observe notification
with exception

Change-Id: I10cf71994872207460e15c1c2446d92f7fac75e8

GitOrigin-RevId: 052d9e4ea5407ebfaf80ed67d6be44a9eff0e1ee
(cherry picked from commit 1184ab659e39fc71e644b6c976ba289ad6fca26b)

GitOrigin-RevId: b28b5ab479c62604b8ed2bab9c7a1acafe448caf