Remove support for className/methodName in navigation
These fields are not provided by the runtime since the early dev
versions of Compose runtime so getting rid of the checks and reflection
calls.

The inline calculation is also not needed anymore since we do not get
the "off file" offsets anymore. The runtime does now provide the correct
line number so no re-mapping needs to be done.

Test: //tools/adt/idea/compose-designer/...
Bug: N/A
Change-Id: I189e391dab4f5ec56b52198dc18bdf9e2ef18e93

GitOrigin-RevId: bb2e2f79203d335516bc9a12b69c91e3b414d7ee