Remove setUpSdks() from AndroidSdk and GradleSpecific initializers
This logic was responsible to configure the SDK and create the
different android targets by populating the jdk.table with those
entries. At the same, this also displayed some custom dialogs asking
the user to specify the SDK in case the wrong one was provided.

In order to simplify the existing implementation we are going to
relay completely on our MissingSdkAlertStep as part of the
Setup wizard that will ask the user to specify a valid SDK path,
otherwise will download any missing components.

Bug: 243036985
Test: Existing
Change-Id: Ibdfd42ff51870c9a3d7898dc9afa483481019d74

GitOrigin-RevId: 5e3c432b597353ca7cb5cdbf0ee94a0c094550c4