Replace uses of GradleCoordinate.parseVersionOnly
Attempt to be clear about whether the thing being parsed is a version
identifier or possibly a rich version specifier, and use the
appropriate parser.  We use the lowerBound even when the original code
uses COMPARE_PLUS_HIGHER because, if there is a prefix specifier, the
app compat migration should in fact replace the prefix with the base
version so that at least that base version is used.

Bug: 279886738
Test: existing
Change-Id: I540748f90b14848598136c3afbad92dd935dcf40

GitOrigin-RevId: 8a7eab2a5a25488c962d278735df1f9d89bfbab3