Change how RenderTask quality is used
This change intends to make the logic around the quality value used
in RenderTask to be simpler and easier to understand. The change
can be summarized as:
- Remove the property 'myMinDownscalingFactor' instead of making the
'quality' to only be applied relative to 1-'myMinDownscalingFactor'
- Make 'myDownScaledImageMaxBytes' be the absolute upper bound of the
image sizes (here a bug around the need of using sqrt is fixed)
- Stop using the calculated downscale factor to move the upper bound
down, but instead always apply it to the image dimensions.

This change also updates the tools using the quality value in
RenderTask to keep the behaviour the same or as similar as possible
to how they work with the previous logic.

Bug: 279604596
Test: RenderTaskTest
Change-Id: I9d2e86b8628235d118aa88725b52f69c6aa3dd01

GitOrigin-RevId: ca06d96c63413151eaea41710f5783d7e8de138f