Add file child node to Compose Visual Lint issues
This adds a child node to the Visual Lint issues coming from Compose.
This node is a file node for the file where the code impacted by the
issue is located. Selecting the node will show the code in the right
panel of the Problems tab. And double-clicking will navigate to the
file at the correct location.

Bug: 274150043
Test: test added
Change-Id: I77fe93a4a0236f387c9c70d8fc602364eb2a12b8

GitOrigin-RevId: a8ed891c52050a4684b0d4e34a7d7955ce1bbf11