Collect coverage in test folders
Unit test execution starts in test folders and from there they exercise
classes in production folders: those classes are the ones we actually
want coverage for. Without this configuration we get nothing in the
Coverage tool window because the coverage runner won't consider
execution that starts in test folders. With it we get the classes we
care about.

Unfortunately we also get coverage for the tests. I haven't figured out
a good way to filter them out. Even so, I feel this is better than
nothing by default.

Bug: Not applicable
Test: Not applicable
Change-Id: Ifa0be85c846eae3d4c322056ef0fa0cf46159811

GitOrigin-RevId: 883e0bd50536b45311c536f728e5bb2ab74459d6