Fix null pointer exception in state chart
When the previousValue variable in the StateChart is null,
because it is of a generic supertype T, it can be null or
not (only known at runtime). However, it is passed into a
function expecting a non-null value. Thus, when passed in,
it is causing a null pointer exception.

To fix this, we can simply check if previousValue is null
or not, and only call the function if it is non-null.

One may say we should fix this by making sure the data is
filtered to not have any null values. Specifically, the
paddedFrames list (located in the AndroidFrameTimelineModel),
which is the data that is causing this exception. However, by
inspecting the call to create the paddedFrames list, one can
see the "pad" parameter in the call to the padding function.
This parameter is set to map two long values into a null value.
Thus, it is clear that the null values being produced is a
purposeful.

Fixes: 285598533
Test: manual
Change-Id: I72066fb6465fa1be421e5a9310ecbae04575823a

GitOrigin-RevId: 4dda4c42f17b41fd7c7cac32076895121ba6af1b