Show active tool window minimize button color based on focus
In the new UI: The WorkBench tool window minimize button will now
be painted bright blue with which text & icon.

Up to now the bright blue background was used whenever the tab is
open. Change this such that the blue color is only used when the
tab is open and the content has focus.

In order to do this we have to keep track of which tool window has
focus. Add this functionality in the global WorkBenchManager.

Fixes: 285617112
Test: Added unit test
Change-Id: I017ea08b9f4f49e26de139cd44df9b2ec15a09be

GitOrigin-RevId: 562f331f3e442c122c178cb67d5ee1c0b3f24a26