TraceboxFlag to load CpuProfiling config for System Trace
If traceboxFlag is enabled and Android is between M and P (both ends
inclusive), then PerfettoConfiguration is assigned and if disabled it
works as it is today (PerfettoConfiguration for >= P else
AtraceConfiguration). PerfettoConfiguration constructor is updated to
accept traceboxFlag as parameter and all required places including tests
are updated to accomodate this change.

Bug: 285412214
Test: Added
Change-Id: Ic0826748dd133629e1779cb414985cfd3fd0cd11

GitOrigin-RevId: ce96a93ce2dc5defaee6dfe1db441c40ca118780