Log cpu (and system) capture and parsing
To implement an e2e test for capturing a system trace,
log statements have been added to identify when the
profiler attempts to start a cpu capture, successfully
starts the capture, attemps to end the cpu capture,
successfully ends the cpu capture, parses the cpu capture,
and idenitifes the system trace data as non-empty.

In this way, all but the last log statement (the one checking
if the system trace is non-empty), can apply to different
types of cpu captures (e.g. call stack) if we want to add
e2e tests for it in the future.

The expected logs (from this CL's additions) should roughly
look like the following when capturing a system trace:

…CpuProfilerStage - PROFILER: CPU capture attempting start
…CpuProfilerStage - PROFILER: CPU capture successfully started
…CpuProfilerStage - PROFILER: CPU capture attempting stop
…$InProgressTraceHandler - PROFILER: CPU capture successfully stopped
…CpuCaptureStage - PROFILER: CPU capture successfully parsed
…CpuCaptureStage - PROFILER: CPU capture contains system trace data

Bug: 270582414
Test: This changes is for end-to-end testing
Change-Id: Idb49bd84d1d6f1f811147b3528bf10b1e6e8e5d3

GitOrigin-RevId: 23d1adc32358d38b14e9e3ddc2d0472bd22202ac