Add an option to skip runtime classpath fetching for libs
This change adds a new checkbox option to GradleExperimentalSettings in
order to allow users to opt-in to skipping the computation of runtime
classpaths where possible for Android library projects.

We re-purpose the existing Studio flag to now act as a control on
whether the option is visible to the user or not. We switch the default
of the Studio flag to true (the option in settings is still off by default).

We adjust usages of the flag to also read from the
GradleExperimentalSettings and only be enabled if BOTH are true. We also
inject the boolean property
android.experimental.dependency.excludeLibraryComponentsFromConstraints
into AGP in order for users to benefit fully from this option.

We also now need to override this additional value in the SyncedProjects
test cases which test this option.

Fixes: 287017263
Test: updated
Change-Id: I2577c7a30d5ecda2d87f6d617012f1e4ff2763b6

GitOrigin-RevId: 71e419f2f45a9aa355ec02f48cc5e8d8d8f1979e