Rearrange "More Actions" on Welcome Screen Project tab
Order them as requested.  Note that the ordering still depends on xml
processing order, in that "Profile or Debug APK" is specified to go
last, as is "SDK Manager".  This is unavoidable with the actions
declared in xml files, given that the separator has no id and so
cannot be referenced as part of the ordering.

Bug: 287418329
Test: manual
Change-Id: I99818b1510daac7753101d0426e65b77bed976dc

GitOrigin-RevId: ff59137237eff752ac7e879883d7a25a289a94cb