Restore original semantics of GradleProjectPath
This change fixes the composite build support after
IDEA 231.8109.175 platform merge. The platform
added support for composite-aware Gradle project
paths, but there is quite a bit of our internal
infra that expects to handle GradleProjectPath. This
abstraction points to a root of a Gradle build (note
this could be an included build), and this change
restores that semantics. Also, GradleProjectPath.path
is non-composite-aware Gradle path, as expected by
users of this API in Studio.

Some callers have been migrated to use the platform
APIs (gradleIdentityPath).

Bug: 283852233
Test: updating them
Change-Id: I97ab2ffe239db379165a8e91f0af6623cd052db8

GitOrigin-RevId: a7d95a632605cb3d38d660ac87dbf140b796b31c