Split-up benchmark names from the project to be used
We add a new argument to the project setup rule, a BenchmarkProject
which defines the base project path, set of diffs to apply and the
maximum heap memory to be used for the project.

Before the projectName was used to determine both the project to be used
and the name of the benchmark to attach the metrics to. With this change
these we decouple these to allow as create tests with different
benchmark names but run off the same project.

This change should be a pure refactoring change and result in no change
in behaviour.

Bug: 287017136
Test: self
Change-Id: I99bbea09b189976cde73284a72539cf1805939a6

GitOrigin-RevId: 0dd9aa4a87138db664e5cb7059416b6dcd1f73c3